+++ src/common/transformers/string.transformer.ts
@@ -3,4 +3,5 @@ import { Transform } from 'class-transformer';
 /**
  * Transform string with trim
  */
-export const TransformTrimString = () => Transform(v => (typeof v === 'string' ? v.trim() : v));
+export const TransformTrimString = () =>
+  Transform(({ value }) => (typeof value === 'string' ? value.trim() : value));