+++ src/lib/pagination/pagination.dto.ts
@@ -11,7 +11,7 @@ export class PaginationDto {
   @ApiPropertyOptional({ default: 10, description: 'Limit query data' })
   @Min(0)
   @Max(100)
-  @Transform(x => x || 1) // Prevented from query limit in mongodb, .limit(0) will return all from db
+  @Transform(x => x.value || 1) // Prevented from query limit in mongodb, .limit(0) will return all from db
   @TransformToNumber() // Don't remove, look the comment above
   readonly limit: number = 10;