+++ src/lib/typeorm/typeorm.dto.ts
@@ -8,7 +8,7 @@ export class TypeOrmConfig {
 
   @IsNotEmpty()
   @IsNumber()
-  @Transform(x => +x)
+  @Transform(x => +x.value)
   DB_PORT!: number;
 
   @IsNotEmpty()
@@ -29,11 +29,11 @@ export class TypeOrmConfig {
 
   @IsNotEmpty()
   @IsBoolean()
-  @Transform((x: string) => x.toLowerCase() === 'true')
+  @Transform(x => String(x.value).toLowerCase() === 'true')
   DB_LOGGING!: boolean;
 
   @IsNotEmpty()
   @IsBoolean()
-  @Transform((x: string) => x.toLowerCase() === 'true')
+  @Transform(x => String(x.value).toLowerCase() === 'true')
   DB_SYNC!: boolean;
 }